requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
pyTelegramBotAPI==3.7.3
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
